HIVE-4673 : Use VectorExpessionWriter to write column vectors into Writables. (Jitendra Nath Pandey via Ashutosh Chauhan)
git-svn-id: https://svn.apache.org/repos/asf/hive/branches/vectorization@1490492 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/io/orc/RecordReaderImpl.java
@@ -552,7 +552,7 @@ class RecordReaderImpl implements RecordReader {
       // Read value entries based on isNull entries
       for (int i = 0; i < batchSize; i++) {
         if (!result.isNull[i]) {
-          result.vector[i] = SerializationUtils.readDouble(stream);
+          result.vector[i] = SerializationUtils.readFloat(stream);
         } else {
 
           // If the value is not present then set NaN
